fix(query): correctly cast embedded discriminator paths when discriminator key is specified in array filter
Fix #9977
