+++ lib/helpers/query/castUpdate.js
@@ -204,7 +204,7 @@ function walkUpdatePath(schema, obj, op, options, context, filter, pref) {
       schematype = schema._getSchema(prefix + key);
 
       if (schematype == null) {
-        const _res = getEmbeddedDiscriminatorPath(schema, obj, filter, prefix + key);
+        const _res = getEmbeddedDiscriminatorPath(schema, obj, filter, prefix + key, options);
         if (_res.schematype != null) {
           schematype = _res.schematype;
         }
@@ -324,7 +324,7 @@ function walkUpdatePath(schema, obj, op, options, context, filter, pref) {
       // If no schema type, check for embedded discriminators because the
       // filter or update may imply an embedded discriminator type. See #8378
       if (schematype == null) {
-        const _res = getEmbeddedDiscriminatorPath(schema, obj, filter, checkPath);
+        const _res = getEmbeddedDiscriminatorPath(schema, obj, filter, checkPath, options);
         if (_res.schematype != null) {
           schematype = _res.schematype;
           pathDetails = _res.type;
