+++ lib/helpers/update/castArrayFilters.js
@@ -3,7 +3,7 @@
 const castFilterPath = require('../query/castFilterPath');
 const cleanPositionalOperators = require('../schema/cleanPositionalOperators');
 const getPath = require('../schema/getPath');
-const modifiedPaths = require('./modifiedPaths');
+const updatedPathsByArrayFilter = require('./updatedPathsByArrayFilter');
 
 module.exports = function castArrayFilters(query) {
   const arrayFilters = query.options.arrayFilters;
@@ -15,24 +15,7 @@ module.exports = function castArrayFilters(query) {
   const schema = query.schema;
   const strictQuery = schema.options.strictQuery;
 
-  const updatedPaths = modifiedPaths(update);
-
-  const updatedPathsByFilter = Object.keys(updatedPaths).reduce((cur, path) => {
-    const matches = path.match(/\$\[[^\]]+\]/g);
-    if (matches == null) {
-      return cur;
-    }
-    for (const match of matches) {
-      const firstMatch = path.indexOf(match);
-      if (firstMatch !== path.lastIndexOf(match)) {
-        throw new Error(`Path '${path}' contains the same array filter multiple times`);
-      }
-      cur[match.substring(2, match.length - 1)] = path.
-        substr(0, firstMatch - 1).
-        replace(/\$\[[^\]]+\]/g, '0');
-    }
-    return cur;
-  }, {});
+  const updatedPathsByFilter = updatedPathsByArrayFilter(update);
 
   for (const filter of arrayFilters) {
     if (filter == null) {
