+++ lib/query.js
@@ -4586,7 +4586,8 @@ Query.prototype._castUpdate = function _castUpdate(obj, overwrite) {
     strict: strict,
     omitUndefined,
     useNestedStrict: useNestedStrict,
-    upsert: upsert
+    upsert: upsert,
+    arrayFilters: this.options.arrayFilters
   }, this, this._conditions);
 };
